Added extended tests
